adding tokenizer methods
