+++ src/be/svlandeg/packeval/nlp/nltk_wrapper.py
@@ -1,22 +1,23 @@
-from nltk import bigrams, trigrams
+import nltk
 
 
 class NLTK:
     @staticmethod
-    def ngrams_split(n, text):
-
+    def ngrams(n, tokens):
         # bigrams using split() on spaces
         if n == 2:
-            result = list(bigrams(text.split()))
+            result = list(nltk.bigrams(tokens))
             return result
 
         # trigrams using split() on spaces
         if n == 3:
-            result = list(trigrams(text.split()))
+            result = list(nltk.trigrams(tokens))
             return result
 
         # default: not supported option returning empty list
         return list()
 
-
-
+    @staticmethod
+    def tokenize_words(text):
+        # nltk.download('punkt')
+        return nltk.word_tokenize(text)
